+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to tree-element are documented here. The format is based on
 
 ## Unreleased
 
+## 0.2.0 - 2026-09-16
+
+### Fixed
+
+- The options `closedIcon` and `openedIcon` render a string as html instead of
+  as text.
+
 ### Changed
 
 - Make rendering faster by using templates.
@@ -13,6 +20,7 @@ All notable changes to tree-element are documented here. The format is based on
   instead of when the tree is rendered. `onCreateLi` is called for a node when
   it is rendered.
 - The method `getSelectedNode` returns null instead of false.
+- The method `setState` saves the state to storage when `saveState` is enabled.
 
 ## 0.1.1 - 2026-09-06